fix: circ import
