+++ a_sync/a_sync/property.py
@@ -4,9 +4,9 @@ import logging
 
 import async_property as ap  # type: ignore [import]
 
-from a_sync import _helpers, _smart, exceptions
+from a_sync import _smart, exceptions
 from a_sync._typing import *
-from a_sync.a_sync import config
+from a_sync.a_sync import _helpers, config
 from a_sync.a_sync._descriptor import ASyncDescriptor
 from a_sync.a_sync.function import ASyncFunction, ASyncFunctionAsyncDefault, ASyncFunctionSyncDefault
 from a_sync.a_sync.method import ASyncBoundMethodAsyncDefault, ASyncMethodDescriptorAsyncDefault
